third_party/blink/web_tests/http/tests/loading/bad-scheme-subframe-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "f1" - BeginNavigation request to 'this-is-definitely-a-bad-uri-scheme:', http method GET
frame "f1" - DidStartNavigation
main frame - didHandleOnloadEventsForFrame
frame "f1" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
